CI: Add workflow to update changelog and rebuild on repo dispatch
This workflow can be triggered using a personal access token (PAT) with
the `public_repo` scope.

    $ curl -H "Accept: application/vnd.github.v3+json" \
           -H "Authorization: token <your-token>" \
           --request POST \
           --data '{"event_type": "update-changelog", "client_payload": {"branch": "2.3"}}' \
           https://api.github.com/repos/mixxxdj/manual/dispatches

It will then run `tools/update_changelog.py` and if any files were
changed, it will commit these changes and then trigger rebuild of the
manual.

For the latter to work, a `MIXXXBOT_CHANGELOG_AUTOUPDATER_PAT`
repository secret needs to be present, because the default
`GITHUB_TOKEN` doesn't suffice.