Pitch slider is no longer opposite to turntable/CDJ behaviour
The manual states that:
> The slider allows you to change the speed of the song, by default
> up to 10% from the tracks original tempos. The speed will increase
> as you move the slider up, opposite to the behavior found on
> DJ turntables and CDJ

This appears to no longer be the case; certainly my Mixxx 2.5.0 wearing
the LateNight skin has the slider increase speed as it goes down - and
this appears the same in the Deere and Tango skins, so it doesn't look
to be skin-specific.

I couldn't find an entry in the changelog to say if/when this changed.