Traktor S3: fix incorrect button description
Previously the VIEW button said that it selected menus and the list-plus
button said that it controlled auto DJ.
However, the list-plus button actually toggles between menus and the
view button maximizes the library view.

This also fixes the description of the rotary encoder to explain that it
does more than just load tracks.

Signed-off-by: Sam Whited <sam@samwhited.com>